+++ src/commands/openUrl.ts
@@ -1,0 +1,5 @@
+import * as vscode from "vscode";
+
+export const openUrl = (url: string) => {
+  vscode.env.openExternal(vscode.Uri.parse(url));
+};